install: Update install/clean scripts
Mostly ensuring that they match what we're currently installing. Special
case for the .clang-format file: it won't be installed automatically,
instead we should use whatever each project has. In the event that we
need a global one, we can always symlink it manually and then the
scripts can remove it if it exists.

+++ clean.sh
@@ -2,25 +2,26 @@
 
 cur_os=$(uname)
 
-# ZSH
+# nvim
+rm "$HOME/.config/nvim"
+
+# git
+rm "$HOME/.gitconfig"
+rm "$HOME/.gitignore"
+
+# zsh
 rm "$HOME/.zshenv"
 rm "$HOME/.config/zsh"
 
-# Git
-rm "$HOME/.gitconfig"
-rm "$HOME/.gitignore"
-rm "$HOME/.gittemplate.txt"
-
-# Nvim
-rm "$HOME/.config/nvim"
-
 # Language files
-rm "$HOME/.clang-format"
+if [ -f "$HOME/.clang-format" ]; then
+    rm "$HOME/.clang-format"
+fi
 
-# Bat
+# bat
 rm "$HOME/.config/bat"
 
-# WezTerm
+# wezterm
 rm "$HOME/.config/wezterm"
 
 if [[ "$cur_os" != "Darwin" ]]; then
